requirements.txt
pyspark==3.5.0
kafka-python
scikit-learn
torch
torchvision
mlflow
pandas
matplotlib
packaging<25